implemented get_chat_name - improved private chans
When a PurpleChat (chat icon in buddy list) wants to get the name of a
channel it's attached to, it tries to call the get_chat_name function
from prpl, and defaults to the conversation name (which is the channel
ID). This works for official channels, where the title matches the ID,
but not for private ones.

This commit adds a function to lookup the channel name from the
'component' data provided when the blist node is created. It makes
tooltips and "Get logs" work properly for private channels, and it
allows private channels to be made persistant ( which fixes #2 )