Remove temporary contacts on startup
In this change, we make the plugin empty its temporary IM list in the
same function as the one that clears the search filter.

Another possibility would be to mark the contact as non-saveable, but in
that case, moving the contact from the temporary group to a permanent
group would prevent it from being saved, I believe this is the same
reason why the search feature does not set this attribute.