Fix #61: Handle password mismatch for ticket requests
When requestion new tickets from the API, a password mismatch will cause
the failed login counter to increment and eventually cause a temporary
block of the user account.

This change allows us to detect password mismatches and terminate the
connection, giving a change to the user to change their passwords in
Pidgin as well.

After changing their password, the user must re-enable their account.