Fix #70: implement client-side limit on local search
This change cowardly refuses to add more than 350 buddies to the buddy
list, as it might crash pidgin, cause it to timeout, and it would be
impossible to browse through anyway.

The number 350 is the same as the server-side limit that is already
enforced when doing searches with a kink filter.

This issue was mentionned first in #68