Remove proper_character variable
A class of errors was identified in #128 as appearing when trying to
access an uninitialized  `proper_character` variable. Since the
`character` variable is hardly ever used to begin with, this commit
merges the two variables. Upon receiving the normalized character name
for the server, it simply replaces the one set in account options