Refactor fonctions to avoid taking PurpleConnection as argument
Many internal functions (not communating with purple) take a
PurpleConnection as an argument. Leading to this pattern all around the
code.

    FListAccount *fla; // comes from function argument usually
    PurpleConnection *pc = fla->pc;
    some_function(pc);

Where some_function starts with :
    FListAccount *fla = pc->proto_data;
and use fla almost exclusively.

This commit reworks the code so that most of our internal functions pass
FListAccounts as argument instead, avoiding unnecessary type gymnastics