Check for message length before sending
Previously, sending a message that was too big resulted in a popup error
message. With this commit, sending a message above server limits will
show an error message in the window. The fact that the message had not
been sent is much clearer to the user, and popups are annoying.